add basic decoding support for PPM and PPT
This decodes most cases of PPT and some cases of PPM without decoding
the packet header. To fully implement decoding of PPM packet header
decoding will need to be implemented to determine if the packet is
complete and therefore the omission of Nppm parameter and therefore the
size of the data.