fix: clippy error / warning fixes
The only functional change is the coding style bit flags (i.e. SPcod) which
were comparing the flag to 1, which is only valid for the selective arithmetic
coding bypass flag. The other flags would need to compare to 2, 4, 8 etc.

There are more clippy warnings, but I'd like to get tests in place before
changing those.