jpc: testing and testability fixes
Derives partial equality for some enums to allow comparison.

Changes the types for some segments in the main header to be Option<...>, to match the optionality
shown in Part 1 Table A.2.

Exposes additional segments in the main header to allow testing and wider use.

Extends existing test to check the values, consistent with the current behaviour.