fix(jp2): update the colourspace handling
This is a general tidy up of the Colour Specification box.

There are two API changes:
 - make the enumerated colour space value return a Rust enum (instead of u32)
 - expose the restricted ICC profile value

We already had the enumeration, just not using it. However one value was
missing from the enum (for sYCC), which is added.

There is a small  behaviour change where we now return None for the enumerated colour space instead
of returning Some (with a reserved value) when the colourspace is method 2 (ICC). I used the same
behaviour when exposing the restricted ICC profile value.

There is a parsing fix for the restricted ICC profile value, where we now store
the parsed value. Previously it was parsed and then dropped.

The rest of the changes are documentation and testing.